corrected many minor warnings using cppcheck..

+++ opengl_acquisition_shared_library/opengl_depth_and_color_renderer/src/Library/MotionCaptureLoader/bvh_to_tri_pose.c
@@ -144,7 +144,7 @@ int parseJointAssociation(
                bvhtri->jointAssociation[jID].rotationOrder[2].rotID=BVH_ROTATION_X;
              break;
              default :
-               fprintf(stderr,"parseJointAssociation: incorrect rotation order %u\n",mc->jointHierarchy[realJID].channelRotationOrder);
+               fprintf(stderr,"parseJointAssociation: incorrect rotation order %u\n",(unsigned int) mc->jointHierarchy[realJID].channelRotationOrder);
              break;
            }
            //-------------------------------------------------------
